+++ tests/src/main/java/testcases/T088_smart_cache_spread_source/Spreadable.java
@@ -1,0 +1,31 @@
+/*
+ * Copyright (c) 2025 Uber Technologies, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package testcases.T088_smart_cache_spread_source;
+
+// Spread source. Each spread method returns a facet holding a reference back to this instance,
+// so the test can prove both facets came from a single construction.
+//
+// Only facetA()/facetB() are public no-arg methods, so only those become spread dependencies.
+public class Spreadable {
+
+    public FacetA facetA() {
+        return new FacetA(this);
+    }
+
+    public FacetB facetB() {
+        return new FacetB(this);
+    }
+}